+++ sandbox/sonar/sonar2/sonarPing/sonarPing.h
@@ -1,6 +1,9 @@
-#include "math/include/Vector3.h"
+//Michael Levashov
+//Robotics@Maryland
+//
 //Used to store a single ping direction vector along with its point number
 
+#include "math/include/Vector3.h"
 #ifndef _SONAR_PING
 #define _SONAR_PING
 
@@ -9,7 +12,7 @@ namespace sonar {
 
 struct sonarPing {
     math::Vector3 direction;
-    double distance;
+    int distance;
     int point_num; //number of points that hydrophone 0 was found at
 };
